# docs/swagger.yaml
definitions:
  handlers.AddResultRequest:
    properties:
      personal_result:
        type: integer
      result_type_id:
        type: integer
      rider_id:
        type: integer
    type: object
...
  handlers.CreateRaceRequest:
    properties:
      date:
        type: string
      duration:
        type: integer
      laps:
        type: integer
      race_type:
        $ref: '#/definitions/models.RaceType'
      time_start:
        type: string
      track_id:
        type: integer
    type: object
...
  handlers.RegisterRiderRequest:
    properties:
      result_type_id:
        type: integer
      rider_id:
        type: integer
    type: object
  handlers.UpdateUserRequest:
    additionalProperties: true
    type: object
...
  models.DifLevel:
    enum:
    - Детский
    - Легкий
    - Средний
    - Сложный
    type: string
    x-enum-varnames:
    - Kids
    - Light
    - Medium
    - Hard
...
  models.Kartodrom:
    properties:
      city:
        type: string
      email:
        type: string
      id:
        type: integer
      latitude:
        type: number
      location:
        type: string
      longitude:
        type: number
      name:
        type: string
      phone:
        type: string
      schedules:
        items:
          $ref: '#/definitions/models.KartodromSchedule'
        type: array
    type: object
  models.KartodromSchedule:
    properties:
      closeTime:
        type: string
      dayOfWeek:
        type: integer
      id:
        type: integer
      kartodrom:
        $ref: '#/definitions/models.Kartodrom'
      kartodromID:
        type: integer
      openTime:
        type: string
    type: object
  models.RaceType:
    enum:
    - Свободный заезд
    - Одиночный заезд на время
    - Парный заезд
    - Заезд на короткую дистанцию
    - Заезд на длинную дистанцию
    type: string
    x-enum-varnames:
    - FreeRide
    - TimeAttack
    - Duo
    - SprintRace
    - EnduranceRace
  models.Track:
    properties:
      difLevel:
        $ref: '#/definitions/models.DifLevel'
      id:
        type: integer
      kartodrom:
        $ref: '#/definitions/models.Kartodrom'
      kartodromID:
        type: integer
      length:
        type: integer
      maxKarts:
        type: integer
      name:
        type: string
      pricePerMin:
        type: number
    type: object
  services.ChangePasswordInput:
    properties:
      new_password:
        type: string
      old_password:
        type: string
      repeat_new_password:
        type: string
    required:
    - new_password
    - old_password
    - repeat_new_password
    type: object
...
info:
  contact: {}
paths:
...
  /kartodrom/{id}/tracks:
    get:
      parameters:
      - description: ID картодрома
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Track'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить доступные треки картодрома
      tags:
      - track
...
  /races:
    post:
      parameters:
      - description: Данные гонки
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateRaceRequest'
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать гонку
      tags:
      - race
  /races/{id}/cancel:
    post:
      parameters:
      - description: ID гонки
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Отменить гонку
      tags:
      - race
  /races/{id}/finish:
    post:
      parameters:
      - description: ID гонки
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Завершить гонку
      tags:
      - race
  /races/{id}/results:
    post:
      parameters:
      - description: ID гонки
        in: path
        name: id
        required: true
        type: integer
      - description: Данные результата
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.AddResultRequest'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Добавить результат участника
      tags:
      - race
  /races/{id}/riders:
    post:
      parameters:
      - description: ID гонки
        in: path
        name: id
        required: true
        type: integer
      - description: Данные участника
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRiderRequest'
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Зарегистрировать участника в гонке
      tags:
      - race
  /races/{id}/riders/{riderId}:
    delete:
      parameters:
      - description: ID гонки
        in: path
        name: id
        required: true
        type: integer
      - description: ID участника
        in: path
        name: riderId
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить участника из гонки
      tags:
      - race
  /races/{id}/start:
    post:
      parameters:
      - description: ID гонки
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Стартовать гонку
      tags:
      - race
...
  /track/{id}:
    get:
      parameters:
      - description: ID трека
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Track'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить трек по ID
      tags:
      - track
  /user/{id}:
    get:
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить профиль пользователя
      tags:
      - user
    patch:
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: Обновляемые данные
        in: body
        name: updates
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateUserRequest'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить профиль пользователя
      tags:
      - user
...
  /user/{id}/change-password:
    post:
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для смены пароля
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/services.ChangePasswordInput'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Сменить пароль пользователя
      tags:
      - user
...
swagger: "2.0"